requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
openai==1.51.0
PyPDF2==3.0.1
python-dotenv
